# backend/internal/handlers/projects.go
package handlers

import (
	"net/http"

	_ "thothix-backend/internal/models"

	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

type ProjectHandler struct {
	db *gorm.DB
}

func NewProjectHandler(db *gorm.DB) *ProjectHandler {
	return &ProjectHandler{db: db}
}

// GetProjects godoc
// @Summary Get all projects
// @Description Get a list of all projects
// @Tags projects
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {array} models.Project
// @Failure 500 {object} map[string]interface{}
// @Router /api/v1/projects [get]
func (h *ProjectHandler) GetProjects(c *gin.Context) {
	// TODO: Implement project listing
	c.JSON(http.StatusOK, gin.H{"message": "Project listing - TODO"})
}

// CreateProject godoc
// @Summary Create a new project
// @Description Create a new project
// @Tags projects
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 201 {object} models.Project
// @Failure 400 {object} map[string]interface{}
// @Router /api/v1/projects [post]
func (h *ProjectHandler) CreateProject(c *gin.Context) {
	// TODO: Implement project creation
	c.JSON(http.StatusCreated, gin.H{"message": "Project creation - TODO"})
}

// GetProject godoc
// @Summary Get project by ID
// @Description Get a single project by its ID
// @Tags projects
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path string true "Project ID"
// @Success 200 {object} models.Project
// @Failure 404 {object} map[string]interface{}
// @Router /api/v1/projects/{id} [get]
func (h *ProjectHandler) GetProject(c *gin.Context) {
	// TODO: Implement get project
	c.JSON(http.StatusOK, gin.H{"message": "Get project - TODO"})
}

// UpdateProject godoc
// @Summary Update project
// @Description Update a project's information
// @Tags projects
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path string true "Project ID"
// @Success 200 {object} models.Project
// @Failure 400 {object} map[string]interface{}
// @Router /api/v1/projects/{id} [put]
func (h *ProjectHandler) UpdateProject(c *gin.Context) {
	// TODO: Implement project update
	c.JSON(http.StatusOK, gin.H{"message": "Project update - TODO"})
}

// DeleteProject godoc
// @Summary Delete project
// @Description Delete a project
// @Tags projects
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path string true "Project ID"
// @Success 204 "No Content"
// @Failure 404 {object} map[string]interface{}
// @Router /api/v1/projects/{id} [delete]
func (h *ProjectHandler) DeleteProject(c *gin.Context) {
	// TODO: Implement project deletion
	c.JSON(http.StatusNoContent, gin.H{})
}

// AddMember godoc
// @Summary Add member to project
// @Description Add a user as a member to a project
// @Tags projects
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path string true "Project ID"
// @Success 200 {object} models.ProjectMember
// @Failure 400 {object} map[string]interface{}
// @Router /api/v1/projects/{id}/members [post]
func (h *ProjectHandler) AddMember(c *gin.Context) {
	// TODO: Implement add member
	c.JSON(http.StatusOK, gin.H{"message": "Add member - TODO"})
}

// RemoveMember godoc
// @Summary Remove member from project
// @Description Remove a user from a project
// @Tags projects
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path string true "Project ID"
// @Param userId path string true "User ID"
// @Success 204 "No Content"
// @Failure 404 {object} map[string]interface{}
// @Router /api/v1/projects/{id}/members/{userId} [delete]
func (h *ProjectHandler) RemoveMember(c *gin.Context) {
	// TODO: Implement remove member
	c.JSON(http.StatusNoContent, gin.H{})
}